исправил ошибку с выделением памяти на размере вектора

+++ include/vector.h
@@ -11,11 +11,11 @@
 typedef struct {
     void **data;
     TypeInfo* typeInfo;   
-    size_t capacity;
+    int capacity;
 } Vector;
 
 
-Vector* createVector(TypeInfo* typeInfo, void *data, size_t capacity, VectorErrors* operationResult);
+Vector* createVector(TypeInfo* typeInfo, void *data, int capacity, VectorErrors* operationResult);
 void free_vector(Vector* v);
 VectorErrors add_vectors(const Vector* v1, const Vector* v2, Vector* result);
 VectorErrors multiply_vectors(const Vector* v1, const Vector* v2, void* result);
